# build: Parallelize building locale