--- a/scripts/i18n-build.js
+++ b/scripts/i18n-build.js
@@ -6,18 +6,28 @@ const {
   i18n: { locales, defaultLocale },
 } = require('../docusaurus.config');
 
+/**
+ *
+ * @param {string} [locale]
+ */
 const updateConfig = async (locale) => {
   const baseUrl = locale !== defaultLocale ? `/${locale}/` : '/';
   const configPath = join(__dirname, '../docusaurus.config.js');
 
   let docusaurusConfig = await fs.readFile(configPath, 'utf-8');
 
-  docusaurusConfig = docusaurusConfig
-    .replace(/baseUrl: '.*?',/, `baseUrl: '${baseUrl}',`);
+  docusaurusConfig = docusaurusConfig.replace(
+    /baseUrl: '.*?',/,
+    `baseUrl: '${baseUrl}',`
+  );
 
   await fs.writeFile(configPath, docusaurusConfig, 'utf-8');
 };
 
+/**
+ *
+ * @param {string} [locale]
+ */
 const processLocale = async (locale) => {
   const start = Date.now();
   const outdir = locale !== defaultLocale ? `--out-dir build/${locale}` : '';
@@ -29,6 +39,22 @@ const processLocale = async (locale) => {
  *
  * @param {string} [locale]
  */
+async function buildLocale(locale) {
+  try {
+    await updateConfig(locale);
+    await processLocale(locale);
+  } catch (e) {
+    // TODO: It will be nice to do some clean up and point to the right file and line
+    console.error(`Locale ${locale} failed. Please check the logs above.`);
+
+    throw e;
+  }
+}
+
+/**
+ *
+ * @param {string | undefined} [locale]
+ */
 const start = async (locale) => {
   const start = Date.now();
 
@@ -37,16 +63,11 @@ const start = async (locale) => {
   console.log('Building the following locales:');
   console.log(localesToBuild);
 
-  for (const locale of localesToBuild) {
-    try {
-      await updateConfig(locale);
-      await processLocale(locale);
-    } catch (e) {
-      // We catch instead of just stopping the process because we want to restore docusaurus.config.js
-      console.error(e);
-      // TODO: It will be nice to do some clean up and point to the right file and line
-      console.error(`Locale ${locale} failed. Please check the logs above.`)
-    }
+  try {
+    await Promise.all(localesToBuild.map(buildLocale));
+  } catch (e) {
+    // We catch instead of just stopping the process because we want to restore docusaurus.config.js
+    console.error(e);
   }
 
   // Restore `docusaurus.config.js` to the default values