Building a locale seems taking a long time in the workflow. So I tried to parallelize the process building locale files. But I'm not sure to resolve the problem by this.

If this change is not needed, I don't mind to close this PR without a notice.