@molant 2022-02-13
Thanks @MikuroXina,

Have you tried running this locally and comparing to the sequential times?
If it's faster, only comment would be to figure out a way to print the right locale if it fails.