@MikuroXina 2022-02-14
Hmm, it looks got worse. Should I close this?

Benchmarked on my laptop with `hyperfine`:

Before:
  Time (mean ± σ):     210.119 s ± 132.326 s    [User: 318.679 s, System: 27.513 s]
  Range (min … max):   166.173 s … 586.647 s    10 runs

After:
  Time (mean ± σ):     270.450 s ± 133.260 s    [User: 1110.374 s, System: 74.988 s]
  Range (min … max):   99.203 s … 437.842 s    10 runs