@molant 2022-02-14
@MikuroXina I believe the hardware where our CI runs is going to be worse than your system so we should probably not parallelize anything. It was worth a shot though, thanks!